# backend_gateway/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PrepSense Gateway API",
...
    "version": "1.4.0",
    "contact": {
      "name": "PrepSense Development Team",
      "email": "dev@prepsense.com",
      "url": "https://github.com/dankimjw/PrepSense"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8001",
      "description": "Development server"
    },
    {
      "url": "http://localhost:8002",
      "description": "Testing server"
    },
    {
      "url": "https://api.prepsense.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/images/upload": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Upload Image",
        "operationId": "upload_image_api_v1_images_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_image_api_v1_images_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "items": {
                      "additionalProperties": true,
                      "type": "object"
                    },
                    "type": "array"
                  },
                  "type": "object",
                  "title": "Response Upload Image Api V1 Images Upload Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/images/save-detected-items": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Save Detected Items",
        "description": "Save detected food items from vision service to the user's pantry in BigQuery.\n\nThis endpoint properly handles the relationship between pantry_items and products tables,\nensuring each item gets unique IDs that are properly incremented.\n\nArgs:\n    request: SaveItemsRequest containing items and user_id\n    pantry_manager: The pantry item manager service\n\nReturns:\n    A status message and details of saved items",
        "operationId": "save_detected_items_api_v1_images_save_detected_items_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveItemsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Save Detected Items Api V1 Images Save Detected Items Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/images/cleanup-detected-items": {
      "delete": {
        "tags": [
          "images"
        ],
        "summary": "Cleanup Detected Items",
        "description": "Cleanup items that were recently added via vision detection.\n\nArgs:\n    request: CleanupRequest containing user_id and optional hours_ago\n    pantry_manager: The pantry item manager service\n\nReturns:\n    A status message and count of deleted items",
        "operationId": "cleanup_detected_items_api_v1_images_cleanup_detected_items_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CleanupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Cleanup Detected Items Api V1 Images Cleanup Detected Items Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/images/debug-items/{user_id}": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "Debug User Items",
        "description": "Debug endpoint to check items and their timestamps.",
        "operationId": "debug_user_items_api_v1_images_debug_items__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Debug User Items Api V1 Images Debug Items  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/images/item/{pantry_item_id}": {
      "delete": {
        "tags": [
          "images"
        ],
        "summary": "Delete Single Item",
        "description": "Delete a single pantry item by its ID.\n\nArgs:\n    pantry_item_id: The ID of the pantry item to delete\n    user_id: The user ID for verification (defaults to 111)\n    pantry_manager: The pantry item manager service\n\nReturns:\n    A status message with deletion result",
        "operationId": "delete_single_item_api_v1_images_item__pantry_item_id__delete",
        "parameters": [
          {
            "name": "pantry_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pantry Item Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 111,
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Single Item Api V1 Images Item  Pantry Item Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/images/scan-items-vision": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Scan Items Vision",
        "description": "Scan items using the original vision service (gpt-4o).\nThis endpoint provides OCR-compatible interface but uses the more accurate vision service.",
        "operationId": "scan_items_vision_api_v1_images_scan_items_vision_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Scan Items Vision Api V1 Images Scan Items Vision Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_image_api_v1_images_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_image_api_v1_images_upload_post"
      },
...
      "CleanupRequest": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "default": 111
          },
          "hours_ago": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hours Ago"
          }
        },
        "type": "object",
        "title": "CleanupRequest"
      },
...
      "DetectedItem": {
        "properties": {
          "item_name": {
            "type": "string",
            "title": "Item Name"
          },
          "quantity_amount": {
            "type": "number",
            "title": "Quantity Amount"
          },
          "quantity_unit": {
            "type": "string",
            "title": "Quantity Unit"
          },
          "expected_expiration": {
            "type": "string",
            "title": "Expected Expiration"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "default": "Uncategorized"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand",
            "default": "Generic"
          }
        },
        "type": "object",
        "required": [
          "item_name",
          "quantity_amount",
          "quantity_unit",
          "expected_expiration"
        ],
        "title": "DetectedItem"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SaveItemsRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DetectedItem"
            },
            "type": "array",
            "title": "Items"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "default": 111
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "SaveItemsRequest"
      },
...
      "ScanRequest": {
        "properties": {
          "image_data": {
            "type": "string",
            "title": "Image Data"
          },
          "mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime Type"
          }
        },
        "type": "object",
        "required": [
          "image_data"
        ],
        "title": "ScanRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token authentication"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key authentication for external services"
      }
    },
...
  },
...
}